Stock price prediction using LSTM:
Lately the amount of machine learning applications has been growing exponentially with the new state-of-the-art models and techniques that have made the impossible, possible.
And one of these application is stock market prediction, so in this article we are going to dive deep into how we could use deep learning and more precisely one of the most famous recurrent neural networks : LSTM, in order to perform stock market prediction. But before we start, it is necessary to mention that stock market prediction is not an easy task since the prediction part could be divided into two : fundamental analysis (sales, earning, profits, …) and technical analysis (historical price, …). Which means numerous factors could affect the stock price trends, but in this tutorial we are going to use only time series forecasting using the historical price of a given stock.

As i said before, we will be using Long Short Term Memory, a commonly used RNN. Considering the type of data that we will be feeding our model and the ability of a RNN to allow information to persist unlike standard feed forward neural networks ( they could only process single data points eg : images ), an LSTM is the best fit for these type of problems.
LSTM could easily process an entire sequence of data and it introduces the memory cell, which make the network able to effectively associate memories and input remote in time.
In this example we will be feeding our model with a set of sequences that will help predict a given price using time steps. So let’s say that we set our time steps to 60, it means that in order to predict the price of a given day we will need to know the prices of 60 days before.